Added jsType
