+++ modules/stores/tree.js
@@ -3,12 +3,11 @@ import {expandTreePath, expandTreeSubpath, getItemByPath, fixPathsInTree} from '
 import {defaultRuleProperties, defaultGroupProperties, defaultOperator, defaultOperatorOptions, defaultRoot} from '../utils/defaultUtils';
 import * as constants from '../constants';
 import uuid from '../utils/uuid';
-import omit from 'lodash/omit';
 import {
-    getFirstOperator, getFuncConfig, getFieldConfig, getOperatorConfig, getFieldWidgetConfig, 
-    getValueSourcesForFieldOp, getOperatorsForField, getWidgetForFieldOp
+    getFirstOperator, getFuncConfig, getFieldConfig, getOperatorsForField
 } from "../utils/configUtils";
-import {defaultValue, deepEqual} from "../utils/stuff";
+import {deepEqual} from "../utils/stuff";
+import {validateValue, getNewValueForFieldOp} from "../utils/validation";
 
 
 const hasChildren = (tree, path) => tree.getIn(expandTreePath(path, 'children1')).size > 0;
@@ -187,282 +186,6 @@ const moveItem = (state, fromPath, toPath, placement, config) => {
     return state;
 };
 
-/**
- * @param {object} config
- * @param {object} oldConfig
- * @param {Immutable.Map} current
- * @param {string} newField
- * @param {string} newOperator
- * @param {string} changedField
- * @return {object} - {canReuseValue, newValue, newValueSrc, newValueType}
- */
-export const getNewValueForFieldOp = function (config, oldConfig = null, current, newField, newOperator, changedField = null, canFix = true) {
-    if (!oldConfig)
-        oldConfig = config;
-    const currentField = current.get('field');
-    const currentOperator = current.get('operator');
-    const currentValue = current.get('value');
-    const currentValueSrc = current.get('valueSrc', new Immutable.List());
-    const currentValueType = current.get('valueType', new Immutable.List());
-
-    const _currentOperatorConfig = getOperatorConfig(oldConfig, currentOperator, currentField);
-    const newOperatorConfig = getOperatorConfig(config, newOperator, newField);
-    const operatorCardinality = newOperator ? defaultValue(newOperatorConfig.cardinality, 1) : null;
-    const currentFieldConfig = getFieldConfig(currentField, oldConfig);
-    const newFieldConfig = getFieldConfig(newField, config);
-
-    // get widgets info
-    const widgetsMeta = Array.from({length: operatorCardinality}, (_ignore, i) => {
-        const vs = currentValueSrc.get(i) || null;
-        const currentWidgets = getWidgetForFieldOp(oldConfig, currentField, currentOperator, vs);
-        const newWidgets = getWidgetForFieldOp(config, newField, newOperator, vs);
-        // need to also check value widgets if we changed operator and current value source was 'field'
-        // cause for select type op '=' requires single value and op 'in' requires array value
-        const currentValueWidgets = vs == 'value' ? currentWidgets : getWidgetForFieldOp(oldConfig, currentField, currentOperator, 'value');
-        const newValueWidgets = vs == 'value' ? newWidgets : getWidgetForFieldOp(config, newField, newOperator, 'value');
-        return {currentWidgets, newWidgets, currentValueWidgets, newValueWidgets};
-    });
-    const currentWidgets = widgetsMeta.map(({currentWidgets}) => currentWidgets);
-    const newWidgets = widgetsMeta.map(({newWidgets}) => newWidgets);
-    const currentValueWidgets = widgetsMeta.map(({currentValueWidgets}) => currentValueWidgets);
-    const newValueWidgets = widgetsMeta.map(({newValueWidgets}) => newValueWidgets);
-    const commonWidgetsCnt = Math.min(newWidgets.length, currentWidgets.length);
-    const reusableWidgets = newValueWidgets.filter(w => currentValueWidgets.includes(w));
-    const firstWidgetConfig = getFieldWidgetConfig(config, newField, newOperator, null, currentValueSrc.first());
-    const valueSources = getValueSourcesForFieldOp(config, newField, newOperator);
-    let canReuseValue = currentField && currentOperator && newOperator 
-        && (!changedField 
-            || changedField == 'field' && !config.settings.clearValueOnChangeField 
-            || changedField == 'operator' && !config.settings.clearValueOnChangeOp)
-        && (currentFieldConfig && newFieldConfig && currentFieldConfig.type == newFieldConfig.type) 
-        && reusableWidgets.length > 0;
-    ;
-    
-    let valueFixes = {};
-    if (canReuseValue) {
-        for (let i = 0 ; i < commonWidgetsCnt ; i++) {
-            const v = currentValue.get(i);
-            const vType = currentValueType.get(i) || null;
-            const vSrc = currentValueSrc.get(i) || null;
-            const isValidSrc = (valueSources.find(v => v == vSrc) != null);
-            const isEndValue = !canFix;
-            const [validateError, fixedValue] = _validateValue(config, newField, newField, newOperator, v, vType, vSrc, canFix, isEndValue);
-            const isValid = !validateError;
-            if (!isValidSrc || !isValid) {
-                canReuseValue = false;
-                break;
-            } else if (canFix && fixedValue !== v) {
-                valueFixes[i] = fixedValue;
-            }
-        }
-    }
-
-    let newValue = null, newValueSrc = null, newValueType = null;
-    newValue = new Immutable.List(Array.from({length: operatorCardinality}, (_ignore, i) => {
-        let v = undefined;
-        if (canReuseValue) {
-            if (i < currentValue.size) {
-                v = currentValue.get(i);
-                if (valueFixes[i] !== undefined) {
-                    v = valueFixes[i];
-                }
-            }
-        } else if (operatorCardinality == 1 && (firstWidgetConfig || newFieldConfig)) {
-            if (newFieldConfig.defaultValue !== undefined)
-                v = newFieldConfig.defaultValue;
-            else if (newFieldConfig.fieldSettings && newFieldConfig.fieldSettings.defaultValue !== undefined)
-                v = newFieldConfig.fieldSettings.defaultValue;
-            else if (firstWidgetConfig.defaultValue !== undefined)
-                v = firstWidgetConfig.defaultValue;
-        }
-        return v;
-    }));
-    newValueSrc = new Immutable.List(Array.from({length: operatorCardinality}, (_ignore, i) => {
-        let vs = null;
-        if (canReuseValue) {
-            if (i < currentValueSrc.size)
-                vs = currentValueSrc.get(i);
-        } else if (valueSources.length == 1) {
-            vs = valueSources[0];
-        } else if (valueSources.length > 1) {
-            vs = valueSources[0];
-        }
-        return vs;
-    }));
-    newValueType = new Immutable.List(Array.from({length: operatorCardinality}, (_ignore, i) => {
-        let vt = null;
-        if (canReuseValue) {
-            if (i < currentValueType.size)
-                vt = currentValueType.get(i);
-        } else if (operatorCardinality == 1 && firstWidgetConfig && firstWidgetConfig.type !== undefined) {
-            vt = firstWidgetConfig.type;
-        }
-        return vt;
-    }));
-
-    return {canReuseValue, newValue, newValueSrc, newValueType};
-};
-
-/**
- * 
- * @param {bool} canFix true is useful for func values to remove bad args
- * @param {bool} isEndValue false if value is in process of editing by user
- * @param {bool} isRawValue false is used only internally from _validateFuncValue
- * @return {array} [validError, fixedValue] - if validError === null and canFix == true, fixedValue can differ from value if was fixed
- */
-const _validateValue = (config, leftField, field, operator, value, valueType, valueSrc, canFix = false, isEndValue = false, isRawValue = true) => {
-    let validError = null;
-    let fixedValue = value;
-
-    if (value != null) {
-        if (valueSrc == 'field') {
-            [validError, fixedValue] = _validateFieldValue(leftField, field, value, valueSrc, valueType, config, operator, isEndValue, canFix);
-        } else if (valueSrc == 'func') {
-            [validError, fixedValue] = _validateFuncValue(leftField, field, value, valueSrc, valueType, config, operator, isEndValue, canFix);
-        } else if (valueSrc == 'value' || !valueSrc) {
-            [validError, fixedValue] = _validateNormalValue(leftField, field, value, valueSrc, valueType, config, operator, isEndValue, canFix);
-        }
-
-        if (!validError) {
-            const fieldConfig = getFieldConfig(field, config);
-            const w = getWidgetForFieldOp(config, field, operator, valueSrc);
-            const fieldWidgetDefinition = omit(getFieldWidgetConfig(config, field, operator, w, valueSrc), ['factory', 'formatValue']);
-            const rightFieldDefinition = (valueSrc == 'field' ? getFieldConfig(value, config) : null);
-    
-            const fn = fieldWidgetDefinition.validateValue;
-            if (typeof fn == 'function') {
-                const args = [
-                    fixedValue, 
-                    //field,
-                    fieldConfig,
-                ];
-                if (valueSrc == 'field')
-                    args.push(rightFieldDefinition);
-                const validResult = fn(...args);
-                if (typeof validResult == "string" || validResult === null) {
-                    validError = validResult;
-                } else {
-                    if (validError == false)
-                        validError = `Invalid value`;
-                }
-            }
-        }
-    }
-
-    if (isRawValue && validError) {
-        validError = `Field ${field}: ${validError}`;
-        console.warn("[RAQB validate]", validError);
-    }
-    
-    return [validError, validError ? value : fixedValue];
-};
-
-/**
- * 
- */
-const _validateNormalValue = (leftField, field, value, valueSrc, valueType, config, operator = null, isEndValue = false, canFix = false) => {
-    let fixedValue = value;
-    const fieldConfig = getFieldConfig(field, config);
-    const w = getWidgetForFieldOp(config, field, operator, valueSrc);
-    const wConfig = config.widgets[w];
-    const wType = wConfig.type;
-
-    if (valueType != wType)
-        return [`Value should have type ${wType}, but got value of type ${valueType}`, value];
-    
-    const fieldSettings = fieldConfig.fieldSettings;
-    if (fieldSettings) {
-        if (fieldSettings.listValues && !fieldSettings.allowCustomValues) {
-            if (value instanceof Array) {
-                for (let v of value) {
-                    if (fieldSettings.listValues[v] == undefined) {
-                        return [`Value ${v} is not in list of values`, value];
-                    }
-                }
-            } else {
-                if (fieldSettings.listValues[value] == undefined) {
-                    return [`Value ${value} is not in list of values`, value];
-                }
-            }
-        }
-        if (fieldSettings.min != null && value < fieldSettings.min) {
-            return [`Value ${value} < min ${fieldSettings.min}`, value];
-        }
-        if (fieldSettings.max != null && value > fieldSettings.max) {
-            return [`Value ${value} > max ${fieldSettings.max}`, value];
-        }
-    }
-    
-    return [null, value];
-};
-
-
-/**
- * 
- */
-const _validateFieldValue = (leftField, field, value, _valueSrc, valueType, config, operator = null, isEndValue = false, canFix = false) => {
-    const {fieldSeparator} = config.settings;
-    const leftFieldStr = Array.isArray(leftField) ? leftField.join(fieldSeparator) : leftField;
-    const rightFieldStr = Array.isArray(value) ? value.join(fieldSeparator) : value;
-    const rightFieldDefinition = getFieldConfig(value, config);
-    if (!rightFieldDefinition)
-        return [`Unknown field ${value}`, value];
-    if (rightFieldStr == leftFieldStr)
-        return [`Can't compare field ${leftField} with itself`, value];
-    if (valueType && valueType != rightFieldDefinition.type)
-        return [`Field ${value} is of type ${rightFieldDefinition.type}, but expected ${valueType}`, value];
-    return [null, value];
-};
-
-/**
- * 
- */
-const _validateFuncValue = (leftField, field, value, _valueSrc, valueType, config, operator = null, isEndValue = false, canFix = false) => {
-    let fixedValue = value;
-    
-    if (value) {
-        const funcKey = value.get('func');
-        if (funcKey) {
-            const funcConfig = getFuncConfig(funcKey, config);
-            if (funcConfig) {
-                if (valueType && funcConfig.returnType != valueType)
-                    return [`Function ${funcKey} should return value of type ${funcConfig.returnType}, but got ${valueType}`, value];
-                for (const argKey in funcConfig.args) {
-                    const argConfig = funcConfig.args[argKey];
-                    const args = fixedValue.get('args');
-                    const argVal = args ? args.get(argKey) : undefined;
-                    const fieldDef = getFieldConfig(argConfig, config);
-                    const argValue = argVal ? argVal.get('value') : undefined;
-                    const argValueSrc = argVal ? argVal.get('valueSrc') : undefined;
-                    if (argValue !== undefined) {
-                        const [argValidError, fixedArgVal] = _validateValue(
-                            config, leftField, fieldDef, operator, argValue, argConfig.type, argValueSrc, canFix, isEndValue, false
-                        );
-                        if (argValidError !== null) {
-                            if (canFix) {
-                                fixedValue = fixedValue.deleteIn(['args', argKey]);
-                                if (argConfig.defaultValue !== undefined) {
-                                    fixedValue = fixedValue.setIn(['args', argKey, 'value'], argConfig.defaultValue);
-                                    fixedValue = fixedValue.setIn(['args', argKey, 'valueSrc'], 'value');
-                                }
-                            } else {
-                                return [`Invalid value of arg ${argKey} for func ${funcKey}: ${argValidError}`, value];
-                            }
-                        } else if (fixedArgVal !== argValue) {
-                            fixedValue = fixedValue.setIn(['args', argKey, 'value'], fixedArgVal);
-                        }
-                    } else if (isEndValue && argConfig.defaultValue === undefined && !canFix) {
-                        return [`Value of arg ${argKey} for func ${funcKey} is required`, value];
-                    }
-                }
-            } else return [`Unknown function ${funcKey}`, value];
-        } // else it's not function value
-    } // empty value
-
-    return [null, fixedValue];
-};
-
-
 
 /**
  * @param {Immutable.Map} state
@@ -567,8 +290,8 @@ const setValue = (state, path, delta, value, valueType, config, __isInternal) =>
 
     const isEndValue = false;
     const canFix = false;
-    const calculatedValueType = valueType || _calculateValueType(value, valueSrc, config);
-    const [validateError, fixedValue] = _validateValue(config, field, field, operator, value, calculatedValueType, valueSrc, canFix, isEndValue);
+    const calculatedValueType = valueType || calculateValueType(value, valueSrc, config);
+    const [validateError, fixedValue] = validateValue(config, field, field, operator, value, calculatedValueType, valueSrc, canFix, isEndValue);
     const isValid = !validateError;
     if (isValid && canFix && fixedValue !== value) {
         value = fixedValue;
@@ -622,7 +345,7 @@ const setOperatorOption = (state, path, name, value) => {
 /**
  * 
  */
-const _calculateValueType = (value, valueSrc, config) => {
+const calculateValueType = (value, valueSrc, config) => {
     let calculatedValueType = null;
     if (value) {
         if (valueSrc === 'field') {
